Fix user ID overflow.

+++ k8asd/Arena/ArenaCommand.cs
@@ -14,7 +14,7 @@ namespace k8asd {
         /// </summary>
         /// <param name="playerId">ID của người chơi bị khiêu chiến.</param>
         /// <param name="playerRank">Thứ hạng hiện tại của người chơi bị khiêu chiến.</param>
-        public static async Task<Packet> DuelArenaAsync(this IPacketWriter writer, int playerId, int playerRank) {
+        public static async Task<Packet> DuelArenaAsync(this IPacketWriter writer, long playerId, int playerRank) {
             return await writer.SendCommandAsync("64007", playerId.ToString(), playerRank.ToString());
         }
     }
