+++ k8asd/Reherse/ReherseInfo.cs
@@ -9,7 +9,7 @@ namespace k8asd
 {
     public class ReherseInfo
     {
-        public int playerid;
+        public long playerid;
         public string playername;
         public int playerlv; //so sao
         public int top;
@@ -17,7 +17,7 @@ namespace k8asd
         public static ReherseInfo Parse(JToken token)
         {
             var result = new ReherseInfo();
-            result.playerid = (int)token["playerid"];
+            result.playerid = (long) token["playerid"];
             result.playername = token["playername"].ToString();
             result.playerlv = (int)token["playerlv"];
             result.top = (int)token["top"];
